cli: refactor attach to the new approach
